+++ src/action/interact/inform.ts
@@ -1,22 +1,33 @@
 import * as ActionFactory from '../../action';
 import { ActionType } from '../../actionType';
+import { AssetTransactionType } from '../../assetTransactionType';
 
-// プロジェクトに変更(2022-05-18~)
 export type IAgent = ActionFactory.IParticipantAsProject;
 export type IRecipient = ActionFactory.IParticipant;
 export type IObject = any;
-export type IPurpose = any;
-export type IResult = any;
-// tslint:disable-next-line:no-empty-interface
-export interface IPotentialActions {
+export interface IPayTransactionAsPurpose {
+    id: string;
+    typeOf: AssetTransactionType.Pay;
 }
-export interface IAttributes<TObject, TResult> extends ActionFactory.IAttributes<ActionType.InformAction, TObject, TResult> {
+export interface IRefundTransactionAsPurpose {
+    id: string;
+    typeOf: AssetTransactionType.Refund;
+}
+export type IPurpose = IPayTransactionAsPurpose | IRefundTransactionAsPurpose;
+export interface IResult {
+    statusCode?: number;
+    useFetchAPI?: boolean;
+}
+export interface IAttributes<TObject> extends Pick<
+    ActionFactory.IAttributes<ActionType.InformAction, TObject, IResult>,
+    'agent' | 'object' | 'project' | 'purpose' | 'recipient' | 'result' | 'typeOf'
+> {
     agent: IAgent;
     recipient: IRecipient;
-    potentialActions?: IPotentialActions;
+    // potentialActions?: IPotentialActions; // discontinue(2024-07-01~)
     purpose?: IPurpose;
 }
 /**
  * 通知アクション
  */
-export type IAction<TAttributes extends IAttributes<IObject, IResult>> = ActionFactory.IAction<TAttributes>;
+export type IAction<TAttributes extends IAttributes<IObject>> = ActionFactory.IAction<TAttributes>;