+++ src/action/trade/pay.ts
@@ -101,7 +101,7 @@ export interface IPaymentService {
     serviceOutput?: IPaymentServiceOutput;
 }
 export type IObject = IPaymentService[];
-export type IInformPayment = IInformActionAttributes<{}, undefined>;
+export type IInformPayment = Pick<IInformActionAttributes<{}>, 'recipient' | 'purpose'>; // optimize(2024-07-01~)
 export interface IPotentialActions {
     add2report: boolean;
     informPayment?: IInformPayment[];