+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - optimize reserve task
 - optimize cancelPendingReservation task
 - optimize reserveTransaction.potentialActions
+- optimize triggerWebhook task
 
 ### Deprecated